+++ tests/integration/test_scripting_engine_popups.gd
@@ -15,7 +15,7 @@ func before_each():
 	if confirm_return is GDScriptFunctionState: # Still working.
 		confirm_return = yield(confirm_return, "completed")
 	cards = draw_test_cards(5)
-	yield(yield_for(0.5), YIELD)
+	yield(yield_for(0.1), YIELD)
 	card = cards[0]
 	target = cards[2]